requirements.txt
# Runtime dependencies
pydantic>=2.10
pydantic-ai>=1.0.0
jsonschema>=4.23
structlog>=24.4.0
typer>=0.15.0
tenacity>=9.0.0
rich>=13.9

# Document processing dependencies
PyMuPDF>=1.25.0
tiktoken>=0.8.0
pytesseract>=0.3.13
pdf2image>=1.17.0
pillow>=11.0.0

# Development dependencies
mypy>=1.14
pytest>=8.0.0
pytest-mock>=3.14.0
pytest-asyncio>=1.0.0
ruff>=0.9.0
